Add CI configuration
This commit adds a basic CI configuration for the qiskit-runtime
repository. It adds 3 jobs, a unittest job that runs the repository's
sole unit test on all supported platforms, a docs build job that
verifies the docs build, and a doc publishing job that will build the
documentation on tags and upload the built html to
https://qiskit.org/documentation/partners/qiskit_runtime . We should add
some lint/style jobs but that can be done in a follow up commit because
it's likely that we'll need to update the code to conform to whichever
tooling we decide on.